Move `MappingFor` function in internal package
