+++ cmd/crank/completion/completion.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 
 	"github.com/posener/complete"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -18,10 +17,8 @@ import (
 	"k8s.io/client-go/restmapper"
 	"k8s.io/client-go/tools/clientcmd"
 	controllerClient "sigs.k8s.io/controller-runtime/pkg/client"
-)
 
-const (
-	errFmtResourceTypeNotFound = "the server doesn't have a resource type %q"
+	"github.com/crossplane/crossplane/cmd/crank/internal"
 )
 
 // Predictors returns all supported predictors.
@@ -117,7 +114,7 @@ func kubernetesResourceNamePredictor() complete.PredictFunc {
 		// If the previously completed argument (resource name) was used by its short form, we need to
 		// get the full resource name to be able to list the resources.
 		rmapper := restmapper.NewShortcutExpander(restmapper.NewDeferredDiscoveryRESTMapper(d), d, nil)
-		mapping, err := mappingFor(rmapper, a.LastCompleted)
+		mapping, err := internal.MappingFor(rmapper, a.LastCompleted)
 		if err != nil {
 			return nil
 		}
@@ -262,44 +259,4 @@ func parseNamespaceOverride(a complete.Args) string {
 		}
 	}
 	return namespace
-}
-
-// Copied over from cli-runtime pkg/resource Builder,
-// https://github.com/kubernetes/cli-runtime/blob/9a91d944dd43186c52e0162e12b151b0e460354a/pkg/resource/builder.go#L768
-func mappingFor(rmapper meta.RESTMapper, resourceOrKindArg string) (*meta.RESTMapping, error) {
-	// TODO(phisco): actually use the Builder.
-	fullySpecifiedGVR, groupResource := schema.ParseResourceArg(resourceOrKindArg)
-	gvk := schema.GroupVersionKind{}
-	if fullySpecifiedGVR != nil {
-		gvk, _ = rmapper.KindFor(*fullySpecifiedGVR)
-	}
-	if gvk.Empty() {
-		gvk, _ = rmapper.KindFor(groupResource.WithVersion(""))
-	}
-	if !gvk.Empty() {
-		return rmapper.RESTMapping(gvk.GroupKind(), gvk.Version)
-	}
-	fullySpecifiedGVK, groupKind := schema.ParseKindArg(resourceOrKindArg)
-	if fullySpecifiedGVK == nil {
-		gvk := groupKind.WithVersion("")
-		fullySpecifiedGVK = &gvk
-	}
-	if !fullySpecifiedGVK.Empty() {
-		if mapping, err := rmapper.RESTMapping(fullySpecifiedGVK.GroupKind(), fullySpecifiedGVK.Version); err == nil {
-			return mapping, nil
-		}
-	}
-	mapping, err := rmapper.RESTMapping(groupKind, gvk.Version)
-	if err != nil {
-		// if we error out here, it is because we could not match a resource or a kind
-		// for the given argument. To maintain consistency with previous behavior,
-		// announce that a resource type could not be found.
-		// if the error is _not_ a *meta.NoKindMatchError, then we had trouble doing discovery,
-		// so we should return the original error since it may help a user diagnose what is actually wrong
-		if meta.IsNoMatchError(err) {
-			return nil, fmt.Errorf(errFmtResourceTypeNotFound, groupResource.Resource)
-		}
-		return nil, err
-	}
-	return mapping, nil
 }
