+++ cmd/crank/internal/client.go
@@ -1,0 +1,54 @@
+// Package internal contains internally used logic for Crossplane CLI.
+package internal
+
+import (
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/api/meta"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+)
+
+const (
+	errFmtResourceTypeNotFound = "the server doesn't have a resource type %q"
+)
+
+// MappingFor returns the RESTMapping for the given resource or kind argument.
+// Copied over from cli-runtime pkg/resource Builder,
+// https://github.com/kubernetes/cli-runtime/blob/9a91d944dd43186c52e0162e12b151b0e460354a/pkg/resource/builder.go#L768
+func MappingFor(rmapper meta.RESTMapper, resourceOrKindArg string) (*meta.RESTMapping, error) {
+	// TODO(phisco): actually use the Builder.
+	fullySpecifiedGVR, groupResource := schema.ParseResourceArg(resourceOrKindArg)
+	gvk := schema.GroupVersionKind{}
+	if fullySpecifiedGVR != nil {
+		gvk, _ = rmapper.KindFor(*fullySpecifiedGVR)
+	}
+	if gvk.Empty() {
+		gvk, _ = rmapper.KindFor(groupResource.WithVersion(""))
+	}
+	if !gvk.Empty() {
+		return rmapper.RESTMapping(gvk.GroupKind(), gvk.Version)
+	}
+	fullySpecifiedGVK, groupKind := schema.ParseKindArg(resourceOrKindArg)
+	if fullySpecifiedGVK == nil {
+		gvk := groupKind.WithVersion("")
+		fullySpecifiedGVK = &gvk
+	}
+	if !fullySpecifiedGVK.Empty() {
+		if mapping, err := rmapper.RESTMapping(fullySpecifiedGVK.GroupKind(), fullySpecifiedGVK.Version); err == nil {
+			return mapping, nil
+		}
+	}
+	mapping, err := rmapper.RESTMapping(groupKind, gvk.Version)
+	if err != nil {
+		// if we error out here, it is because we could not match a resource or a kind
+		// for the given argument. To maintain consistency with previous behavior,
+		// announce that a resource type could not be found.
+		// if the error is _not_ a *meta.NoKindMatchError, then we had trouble doing discovery,
+		// so we should return the original error since it may help a user diagnose what is actually wrong
+		if meta.IsNoMatchError(err) {
+			return nil, fmt.Errorf(errFmtResourceTypeNotFound, groupResource.Resource)
+		}
+		return nil, err
+	}
+	return mapping, nil
+}
